Got binding errors working...
A few things learned here... big one is that the
binding errors will only be found if the incoming
BR is tied name-wise to the model that's placed
in the ModelAndView. So, for example... I usually
place "model" => new HomeModel in the ModelAndView
and  Spring will infer a name based on the type of
the incoming parameter (say homeModel) It'll
then lazily create a BindingResult (and do no
actual binding?!) and show no errors. So... to
fix this make sure the BR is given the right
name (in my case this means using @ModelAttribute
to change the name to "model")